+++ package/kedro_viz/api/rest/responses.py
@@ -3,11 +3,8 @@
 import abc
 from typing import Any, Dict, List, Optional, Union
 
-import fsspec
 import orjson
-from fastapi.encoders import jsonable_encoder
 from fastapi.responses import JSONResponse, ORJSONResponse
-from kedro.io.core import get_protocol_and_path
 from pydantic import BaseModel
 
 from kedro_viz.data_access import data_access_manager
@@ -29,6 +26,10 @@ class APIErrorMessage(BaseModel):
 class BaseAPIResponse(BaseModel, abc.ABC):
     class Config:
         orm_mode = True
+
+class UserCredentials(BaseModel):
+    awsRegion: str
+    bucketName: str
 
 
 class BaseGraphNodeAPIResponse(BaseAPIResponse):
@@ -350,71 +351,3 @@ def get_selected_pipeline_response(registered_pipeline_id: str):
         selected_pipeline=registered_pipeline_id,
         modular_pipelines=modular_pipelines_tree,
     )
-
-
-def save_api_responses_to_fs(filepath: str):
-    protocol, path = get_protocol_and_path(filepath)
-    remote_fs = fsspec.filesystem(protocol)
-    
-    def encode_response(response):
-        jsonable_response = jsonable_encoder(response)
-        return EnhancedORJSONResponse.encode_to_human_readable(jsonable_response)
-
-    def write_to_file(location, data):
-        with remote_fs.open(location, "wb") as f:
-            f.write(data)
-
-    if protocol == "file":
-        for loc in [path, f"{path}/api/nodes", f"{path}/api/pipelines"]:
-            remote_fs.makedirs(loc, exist_ok=True)
-
-    default_response = get_default_response()
-    write_to_file(f"{path}/api/main", encode_response(default_response))
-
-    for node in data_access_manager.nodes.get_node_ids():
-        node_response = get_node_metadata_response(node)
-        write_to_file(f"{path}/api/nodes/{node}", encode_response(node_response))
-
-    for pipeline in data_access_manager.registered_pipelines.get_pipeline_ids():
-        pipeline_response = get_selected_pipeline_response(pipeline)
-        write_to_file(f"{path}/api/pipelines/{pipeline}", encode_response(pipeline_response))
-
-
-def save_api_responses_to_fs(filepath: str):
-    protocol, path = get_protocol_and_path(filepath)
-    remote_fs = fsspec.filesystem(protocol)
-    default_response = get_default_response()
-    jsonable_default_response = jsonable_encoder(default_response)
-    encoded_response = EnhancedORJSONResponse.encode_to_human_readable(
-        jsonable_default_response
-    )
-
-    main_loc = f"{path}/api/main"
-    nodes_loc = f"{path}/api/nodes"
-    pipelines_loc = f"{path}/api/pipelines"
-
-    if protocol == "file":
-        remote_fs.makedirs(path, exist_ok=True)
-        remote_fs.makedirs(nodes_loc, exist_ok=True)
-        remote_fs.makedirs(pipelines_loc, exist_ok=True)
-
-    with remote_fs.open(main_loc, "wb") as f:
-        f.write(encoded_response)
-
-    for node in data_access_manager.nodes.get_node_ids():
-        node_response = get_node_metadata_response(node)
-        jsonable_node_response = jsonable_encoder(node_response)
-        encoded_response = EnhancedORJSONResponse.encode_to_human_readable(
-            jsonable_node_response
-        )
-        with remote_fs.open(f"{nodes_loc}/{node}", "wb") as f:
-            f.write(encoded_response)
-
-    for pipeline in data_access_manager.registered_pipelines.get_pipeline_ids():
-        pipeline_response = get_selected_pipeline_response(pipeline)
-        jsonable_pipeline_response = jsonable_encoder(pipeline_response)
-        encoded_response = EnhancedORJSONResponse.encode_to_human_readable(
-            jsonable_pipeline_response
-        )
-        with remote_fs.open(f"{pipelines_loc}/{pipeline}", "wb") as f:
-            f.write(encoded_response)
